Add docs for Modifier.cloudy and refactor packages
