--- cloudy/src/main/kotlin/com/skydoves/cloudy/internals/ContextExtensions.kt
@@ -1,33 +1,0 @@
-/*
- * Designed and developed by 2022 skydoves (Jaewoong Eum)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.skydoves.cloudy.internals
-
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
-
-/** Returns an activity from Context. */
-@JvmSynthetic
-internal fun Context.getActivity(): Activity? {
-  var context = this
-  while (context is ContextWrapper) {
-    if (context is Activity) {
-      return context
-    }
-    context = context.baseContext
-  }
-  return null
-}
